requirement.txt
Django==2.2.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
djangorestframework==3.12.1
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.4.1
